fix: migrated to new plugin
