+++ .github/workflows/release.yml
@@ -1,4 +1,4 @@
-name: Release to PyPI
+name: Semantic Release
 
 on:
   push:
@@ -6,34 +6,109 @@ on:
       - main
 
 permissions:
-  contents: write
-  id-token: write
+  contents: read
 
 jobs:
   release:
     runs-on: ubuntu-latest
+    outputs:
+      released: ${{ steps.release.outputs.released }}
+      tag: ${{ steps.release.outputs.tag }}
+
+    concurrency:
+      group: ${{ github.workflow }}-release-${{ github.ref_name }}
+      cancel-in-progress: false
+
+    permissions:
+      contents: write
+      issues: write
+      pull-requests: write
+
     steps:
-      - name: Create release and publish to PyPI
-        uses: raven-wing/ez-python-release@initial
+      - name: Generate bot app token
+        id: generate_token
+        uses: actions/create-github-app-token@29824e69f54612133e76f7eaac726eef6c875baf
+        with:
+          app-id: ${{ secrets.MY_RELEASER_ID }}
+          private-key: ${{ secrets.MY_RELEASER_PRIVATE_KEY }}
+
+      - name: Checkout
+        uses: actions/checkout@v6
+        with:
+          token: ${{ steps.generate_token.outputs.token }}
+          fetch-depth: 0
+
+      - name: Set up Python 3.10
+        uses: actions/setup-python@v6
         with:
           python-version: "3.10"
-          github_app_id: ${{ secrets.MY_RELEASER_ID }}
-          github_app_private_key: ${{ secrets.MY_RELEASER_PRIVATE_KEY }}
+          cache: 'pip'
 
-  publish-to-pypi:
-    environment:
-      name: pypi
-      url: https://pypi.org/p/platzky-msgbar
+      - name: Cache Poetry virtualenv
+        uses: actions/cache@v5
+        with:
+          path: |
+            ~/.cache/pypoetry/virtualenvs
+            ~/.cache/pip
+          key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-poetry-
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install poetry==2.0.1
+          poetry install
+
+      - name: Semantic Version Release
+        id: release
+        uses: python-semantic-release/python-semantic-release@350c48fcb3ffcdfd2e0a235206bc2ecea6b69df0
+        with:
+          github_token: ${{ steps.generate_token.outputs.token }}
+          git_committer_name: "semantic-release"
+          git_committer_email: "semantic-release@users.noreply.github.com"
+          build: false
+
+      - name: Build package
+        if: steps.release.outputs.released == 'true'
+        run: poetry build
+
+      - name: Upload to GitHub Release Assets
+        uses: python-semantic-release/publish-action@310a9983a0ae878b29f3aac778d7c77c1db27378
+        if: steps.release.outputs.released == 'true'
+        with:
+          github_token: ${{ steps.generate_token.outputs.token }}
+          tag: ${{ steps.release.outputs.tag }}
+
+      - name: Upload Distribution Artifacts
+        uses: actions/upload-artifact@v6
+        if: steps.release.outputs.released == 'true'
+        with:
+          name: distribution-artifacts
+          path: dist
+          if-no-files-found: error
+
+  deploy:
     runs-on: ubuntu-latest
     needs: release
+    if: ${{ needs.release.outputs.released == 'true' }}
+
+    environment:
+      name: pypi
+      url: https://pypi.org/p/platzky_msgbar
+
     permissions:
+      contents: read
       id-token: write
+
     steps:
-      - uses: actions/download-artifact@v4
+      - name: Download Build Artifacts
+        uses: actions/download-artifact@v7
         with:
-          name: python-package-distributions
-          path: dist/
-      - name: Publish package distributions to PyPI
+          name: distribution-artifacts
+          path: dist
+
+      - name: Publish package to PyPI
         uses: pypa/gh-action-pypi-publish@ed0c53931b1dc9bd32cbe73a98c7f6766f8a527e
         with:
           packages-dir: dist
