+++ .github/workflows/tests.yml
@@ -4,51 +4,120 @@ on:
   pull_request:
     branches:
       - main
-      - feature/plugin-with-tests
+    paths-ignore:
+      - '**.md'
+      - 'docs/**'
+      - 'LICENSE'
+      - '.gitignore'
+      - 'CHANGELOG.md'
   push:
     branches:
       - main
-      - feature/plugin-with-tests
+    paths-ignore:
+      - '**.md'
+      - 'docs/**'
+      - 'LICENSE'
+      - '.gitignore'
+      - 'CHANGELOG.md'
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
 
 permissions:
   contents: 'read'
 
 jobs:
-  tests:
+  lint:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v6
     - name: Set up Python 3.10
-      uses: actions/setup-python@v5
+      uses: actions/setup-python@v6
       with:
         python-version: "3.10"
+        cache: 'pip'
+    - name: Cache Poetry virtualenv
+      uses: actions/cache@v5
+      with:
+        path: |
+          ~/.cache/pypoetry/virtualenvs
+          ~/.cache/pip
+        key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
+        restore-keys: |
+          ${{ runner.os }}-poetry-
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install poetry==1.7.1
+        pip install poetry==2.0.1
         poetry install
     - name: Run linters
       run: make lint-check
 
-    - name: Run unit-tests
-      run: make unit-tests
+  test:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v6
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v6
+      with:
+        python-version: "3.10"
+        cache: 'pip'
+    - name: Cache Poetry virtualenv
+      uses: actions/cache@v5
+      with:
+        path: |
+          ~/.cache/pypoetry/virtualenvs
+          ~/.cache/pip
+        key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
+        restore-keys: |
+          ${{ runner.os }}-poetry-
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry==2.0.1
+        poetry install
+    - name: Run tests with coverage
+      run: make coverage
+    - name: Coveralls
+      uses: coverallsapp/github-action@v2
+      with:
+        path-to-lcov: "./coverage.lcov"
+        github-token: ${{ secrets.GITHUB_TOKEN }}
 
+  e2e:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v6
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v6
+      with:
+        python-version: "3.10"
+        cache: 'pip'
+    - name: Cache Poetry virtualenv
+      uses: actions/cache@v5
+      with:
+        path: |
+          ~/.cache/pypoetry/virtualenvs
+          ~/.cache/pip
+        key: ${{ runner.os }}-poetry-${{ hashFiles('**/poetry.lock') }}
+        restore-keys: |
+          ${{ runner.os }}-poetry-
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install poetry==2.0.1
+        poetry install
     - name: Set up Node.js
       uses: actions/setup-node@v4
       with:
         node-version: '18'
-
     - name: Install Cypress
       working-directory: tests/e2e_tests
       run: |
         npm install cypress
-
     - name: Run e2e tests
       run: |
         poetry run flask --app "tests.e2e_tests.e2e_app:create_app(config_path='tests/e2e_tests/e2e_test_config.yml')" run --debug --host=localhost &
         sleep 6
         cd tests/e2e_tests && npx cypress run --spec "cypress/e2e/**/*"
-
-
-
-
