+++ .github/workflows/todo.yml
@@ -3,12 +3,11 @@ on:
   push:
     branches:
       - main
-      - feature/plugin-with-tests
 
 jobs:
   build:
     runs-on: "ubuntu-latest"
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v6
     - name: "TODO to Issue"
-      uses: "alstr/todo-to-issue-action@v4"
+      uses: "alstr/todo-to-issue-action@v5"
