+++ pyproject.toml
@@ -8,6 +8,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
+platzky = "^1.4.2"
 pydantic = "^2.10.3"
 pydantic-extra-types = "^2.10.1"
 markdown = "^3.10"
@@ -19,20 +20,89 @@ black = "^24.10.0"
 ruff = "^0.8.3"
 pyright = "^1.1.390"
 coverage = "^7.6.9"
-platzky = "^0.2.13"
 interrogate = "^1.7.0"
+python-semantic-release = "^10.4.1"
+
+[tool.black]
+line-length = 100
+target-version = ["py310"]
+
+[tool.ruff]
+line-length = 100
+target-version = "py310"
+show-fixes = true
+
+[tool.ruff.lint]
+select = [
+    "I",
+    "F",
+    "E",
+    "W",
+    "RUF",
+]
+
+[tool.pyright]
+pythonVersion = "3.10"
+pythonPlatform = "All"
+typeCheckingMode = "strict"
+reportMissingImports = true
+reportMissingTypeStubs = false
+reportMissingParameterType = false
+reportUnknownArgumentType = false
+reportUnknownMemberType = false
+reportUnknownVariableType = false
+reportUnknownParameterType = false
+reportUnnecessaryTypeIgnoreComment = true
+reportUntypedClassDecorator = false
+reportUntypedFunctionDecorator = false
+reportUnnecessaryIsInstance = false
+reportUnusedFunction = false
+reportUntypedBaseClass = false
+
+[tool.pytest.ini_options]
+markers = [
+    "skip_coverage: skip coverage for this test",
+]
+
+[tool.interrogate]
+fail-under = 95
+ignore-init-method = true
+ignore-init-module = true
+ignore-magic = true
+ignore-private = true
+ignore-semiprivate = false
+exclude = ["tests", "docs"]
 
 [tool.semantic_release]
 version_toml = ["pyproject.toml:tool.poetry.version"]
-branch = "main"
-build_command = "pip install poetry && poetry build"
-major_on_zero = false
+build_command = "poetry build"
+dist_path = "dist/"
 upload_to_release = true
+remove_dist = false
+commit_author = "semantic-release <semantic-release>"
+commit_message = "chore(release): {version} [skip ci]"
+tag_format = "{version}"
+
+[tool.semantic_release.changelog]
+exclude_commit_patterns = []
+
+[tool.semantic_release.changelog.default_templates]
+changelog_file = "CHANGELOG.md"
+output_format = "md"
+mask_initial_release = true
+
+[tool.semantic_release.branches.main]
+match = "main"
+prerelease = false
 
 [tool.semantic_release.commit_parser_options]
-allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
+allowed_tags = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "build", "ci", "chore", "revert"]
 minor_tags = ["feat"]
 patch_tags = ["fix", "perf"]
+
+[tool.semantic_release.publish]
+dist_glob_patterns = ["dist/*"]
+upload_to_vcs_release = true
 
 [build-system]
 requires = ["poetry-core"]
